fix lint issue, move test
